added the modal files so that the code is cleaner and the vidoe frames can be faster upload
